+++ main.cpp
@@ -11,6 +11,7 @@
 #include "Matrix33.h"
 #include "Ray.h"
 #include "Material.h"
+#include "Camera.h"
 
 using namespace std;
 
@@ -85,37 +86,37 @@ class Light {
         Light(vec3 p, vec3 col) : pos(p), color(col) { }
 };
 
-class Camera {
-    public:
-        int width, height;
-        double scale;
-        vec3 pos;
-        Matrix33 rot;
+// class Camera {
+//     public:
+//         int width, height;
+//         double scale;
+//         vec3 pos;
+//         Matrix33 rot;
 
-        Camera(double fovy_, int width_, int height_) :
-               scale(tan(fovy_)), width(width_), height(height_) {
-        }
+//         Camera(double fovy_, int width_, int height_) :
+//                scale(tan(fovy_)), width(width_), height(height_) {
+//         }
 
-        //Ported for Josh's camera code
-        void lookAt(vec3 target, double rho, double phi, double theta) {
-            vec3 yStd(0.0, 1.0, 0.0), zStd(0.0, 0.0, 1.0);
-            vec3 z = vec3::Spherical(1.0, phi, theta);
-            vec3 y = vec3::Spherical(1.0, M_PI / 2.0 - phi, theta + M_PI);
+//         //Ported for Josh's camera code
+//         void lookAt(vec3 target, double rho, double phi, double theta) {
+//             vec3 yStd(0.0, 1.0, 0.0), zStd(0.0, 0.0, 1.0);
+//             vec3 z = vec3::Spherical(1.0, phi, theta);
+//             vec3 y = vec3::Spherical(1.0, M_PI / 2.0 - phi, theta + M_PI);
 
-            this->rot = Matrix33::BasisRotation(yStd, zStd, y, z);
-            this->pos = z * rho + target;
-        }
+//             this->rot = Matrix33::BasisRotation(yStd, zStd, y, z);
+//             this->pos = z * rho + target;
+//         }
 
-        Ray getRay(double screen_x, double screen_y) {
-            double x = (2 * screen_x / (double) width - 1) * scale;
-            double y = (1 - 2 * screen_y / (double) height) * scale;
+//         Ray getRay(double screen_x, double screen_y) {
+//             double x = (2 * screen_x / (double) width - 1) * scale;
+//             double y = (1 - 2 * screen_y / (double) height) * scale;
 
-            vec3 dir = this->rot * vec3(x, y, -1);
-            dir.normalize();
+//             vec3 dir = this->rot * vec3(x, y, -1);
+//             dir.normalize();
 
-            return Ray(this->pos, dir);
-        }
-};
+//             return Ray(this->pos, dir);
+//         }
+// };
 
 
 Intersection findIntersect(const Ray &ray, vector<Shape *> &shapes) {
@@ -208,8 +209,8 @@ vec3 trace(const Ray &ray, const vec3 &cameraPos,
 
 // TODO: ADD COMMAND LINE OPTIONS
 int main() {
-    const int WIDTH = 512;
-    const int HEIGHT = 512;
+    const int WIDTH = 1024;
+    const int HEIGHT = 1024;
 
     cimg_library::CImg<double> img(WIDTH, HEIGHT, 1, 3);
     img.fill(0.0);
